+++ packages/melody-parser/src/Parser.js
@@ -25,6 +25,8 @@ import {
     copyLoc,
     createNode,
 } from './util';
+import { GenericTagParser } from './GenericTagParser';
+import { createMultiTagParser } from './GenericMultiTagParser';
 import { voidElements } from './elementInfo';
 import * as he from 'he';
 
@@ -61,9 +63,15 @@ export default class Parser {
                 ignoreDeclarations: true,
                 decodeEntities: true,
                 preserveSourceLiterally: false,
+                allowUnknownTags: false,
+                multiTags: {}, // e.g. { "nav": ["endnav"], "switch": ["case", "default", "endswitch"]}
             },
             options
         );
+        // If there are custom multi tags, then we allow all custom tags
+        if (Object.keys(this.options.multiTags).length > 0) {
+            this.options.allowUnknownTags = true;
+        }
     }
 
     applyExtension(ext) {
@@ -434,36 +442,60 @@ export default class Parser {
         }
     }
 
-    error(options) {
-        this.tokens.error(options.title, options.pos, options.advice);
+    error(options, metadata = {}) {
+        this.tokens.error(
+            options.title,
+            options.pos,
+            options.advice,
+            1,
+            metadata
+        );
+    }
+
+    getGenericParserFor(tagName) {
+        if (this.options.multiTags[tagName]) {
+            return createMultiTagParser(
+                tagName,
+                this.options.multiTags[tagName]
+            );
+        } else {
+            return GenericTagParser;
+        }
     }
 
     matchTag() {
         const tokens = this.tokens;
-        const tagStartToken = tokens.la(-1),
-            tagNameToken = tokens.la(0);
+        const tagStartToken = tokens.la(-1);
 
-        const tag = tokens.expect(Types.SYMBOL),
-            parser = this[TAG][tag.text];
+        const tag = tokens.expect(Types.SYMBOL);
+        let parser = this[TAG][tag.text];
+        let isUsingGenericParser = false;
         if (!parser) {
-            tokens.error(
-                `Unknown tag "${tag.text}"`,
-                tag.pos,
-                `Expected a known tag such as\n- ${Object.getOwnPropertyNames(
-                    this[TAG]
-                ).join('\n- ')}`,
-                tag.length
-            );
+            if (this.options.allowUnknownTags) {
+                parser = this.getGenericParserFor(tag.text);
+                isUsingGenericParser = true;
+            } else {
+                tokens.error(
+                    `Unknown tag "${tag.text}"`,
+                    tag.pos,
+                    `Expected a known tag such as\n- ${Object.getOwnPropertyNames(
+                        this[TAG]
+                    ).join('\n- ')}`,
+                    tag.length
+                );
+            }
         }
 
         const result = parser.parse(this, tag);
         const tagEndToken = tokens.la(-1);
-        result.trimLeft = tagStartToken.text.endsWith('-');
-        result.trimRight = tagEndToken.text.startsWith('-');
+        if (!isUsingGenericParser) {
+            result.trimLeft = tagStartToken.text.endsWith('-');
+            result.trimRight = tagEndToken.text.startsWith('-');
+        }
 
         setStartFromToken(result, tagStartToken);
         setEndFromToken(result, tagEndToken);
-        setMarkFromToken(result, 'tagNameLoc', tagNameToken);
+        setMarkFromToken(result, 'tagNameLoc', tag);
 
         return result;
     }
@@ -588,15 +620,22 @@ export default class Parser {
                 } else if (token.type === Types.LBRACKET) {
                     node = this.matchMap();
                 } else {
-                    this.error({
-                        title:
-                            'Unexpected token "' +
-                            token.type +
-                            '" of value "' +
-                            token.text +
-                            '"',
-                        pos: token.pos,
-                    });
+                    this.error(
+                        {
+                            title:
+                                'Unexpected token "' +
+                                token.type +
+                                '" of value "' +
+                                token.text +
+                                '"',
+                            pos: token.pos,
+                        },
+                        {
+                            errorType: 'UNEXPECTED_TOKEN',
+                            tokenText: token.text,
+                            tokenType: token.type,
+                        }
+                    );
                 }
                 break;
         }